+++ Sources/Input/ControlKeySide.swift
@@ -1,4 +1,0 @@
-nonisolated enum ControlKeySide: Equatable {
-    case left
-    case right
-}